Validacion de pertenencia de contactos

+++ delete.php
@@ -4,14 +4,14 @@ require "database.php";
 
 session_start();
 
-if(!isset($_SESSION["user"])){
+if (!isset($_SESSION["user"])) {
   header("Location: login.php");
   return;
 }
 
 $id = $_GET["id"];
 
-$statement = $conn->prepare("SELECT * FROM contacts WHERE id = :id");
+$statement = $conn->prepare("SELECT * FROM contacts WHERE id = :id LIMIT 1");
 $statement->execute([":id" => $id]);
 
 if ($statement->rowCount() == 0) {
@@ -19,7 +19,16 @@ if ($statement->rowCount() == 0) {
   echo ("HTTP 404 NOT FOUND");
   return;
 }
+
+$contact = $statement->fetch(PDO::FETCH_ASSOC);
+
+if ($contact["user_id"] !== $_SESSION["user"]["id"]) {
+  http_response_code(403);
+  echo ("HTTP 403 NOT AUTHORIZED");
+  return;
+}
 $conn->prepare("DELETE FROM contacts WHERE id = :id")->execute([":id" => $id]);
+
 
 
 header("Location:home.php");
